Implement show bucket '/bucketlist/:id'
Why?
To get a articular bucketlist
The API consumer can make a get request to /bucketlist/:id
And get the bucketlist with the specified ID

How?
Define method 'show' in the BucketlistController
Set the method to find the bucketlist with the
Return the bucketlist if found
Or return custome 'Not Found' error message

Completing https://www.pivotaltracker.com/story/show/111717950
[finished #111717950]