Implement Pagination
Why?
The API user should  be able to make a call to the API and request a
particular page and number of records

How?
Add a scope `paginate` to the Bucketlist model to apply the `limit` and
`offset` calls to the returned bucketlists
Append the `.paginate` method call to the bucketlists call in the
Bucketlist#index

Completing https://www.pivotaltracker.com/story/show/111833397
[finished #111833397]