Implement Search By Name
Why?
An API consumer should be able to make a call to the API with a `q` URL
param and it will return the bucketlist with name matching the search string

How?
Add a scope `search` to the Bucketlist model to return bucketlists that
contain the given string in their name
Append the call to the `search` in the `Bucketlists#index`

Completing https://www.pivotaltracker.com/story/show/111833413
[finished #111833413]