Implement Review Corrections
Why?
Some inefficiencies were pointed out in the code during the defence
that needs to be fixed

How?
Refactor the BucketlistsController and ItemsController and the models to
enhance the efficiency of the code by minimizing the number of commands that needs to by run.

Completing https://www.pivotaltracker.com/story/show/112498505
[finished #112498505]